Minor adaptation to accept JORF database
With this patch legi.py can now read the JORF database, which
has a structure very similar to LEGI. The main difference is
the directory structure: articles are not in a subdirectory of
the text but grouped together in a common directory.

To avoid breaking back-compatibility the JORF must be put in
a different SQLite database, and an additional db_meta parameter
is added (called “base”, which can have values in ['JORF', 'LEGI'].

Given the CID is not known in the path itself but only in the file,
the file must be read a bit earlier.